requirements.txt
python-dotenv==1.0.1
Jinja2==3.1.4
pandas==2.2.2
